+++ src/plugin/importCSSVariables/index.ts
@@ -1,0 +1,8 @@
+const importCSSVariables = async ({collectionName, textareaValue}:{
+  collectionName: any,
+  textareaValue: any
+}) => {
+  console.log("importing CSS variables", collectionName, textareaValue)
+}
+
+export default importCSSVariables;
